+++ pkg/client/cli/cloud_messages_test.go
@@ -16,6 +16,12 @@ func newTestContext(t *testing.T) context.Context {
 	// Create a fake user cache directory
 	ctx = filelocation.WithUserHomeDir(ctx, t.TempDir())
 
+	env, err := client.LoadEnv(ctx)
+	if err != nil {
+		t.Error(err)
+	}
+	ctx = client.WithEnv(ctx, env)
+
 	// Load config (will be default since home dir is fake)
 	cfg, err := client.LoadConfig(ctx)
 	if err != nil {
@@ -126,10 +132,7 @@ func Test_cloudUpdateMessages(t *testing.T) {
 }
 
 func Test_cloudRefreshMessagesConfig(t *testing.T) {
-	ctx := dlog.NewTestContext(t, false)
-
-	// Create a fake user cache directory + config directory
-	ctx = filelocation.WithUserHomeDir(ctx, t.TempDir())
+	ctx := newTestContext(t)
 	confDir := t.TempDir()
 
 	// Update the config to a shorter time